[frontend] add a validation dialog on closing form
