+++ portal-front/messages/fr.json
@@ -100,6 +100,7 @@
     "SubscribeOrganization": "Souscrire une organisation",
     "GoTo": "Aller au service",
     "GoToLabel": "Consulter",
+    "GoToManagement": "Aller à la gestion des services",
     "GoToAdminLabel": "Administrer",
     "Capabilities": {
       "CapabilitiesTitle": "Capacités",
@@ -115,6 +116,7 @@
       "LastName": "Nom",
       "OneUserMax": "1 utilisateur max",
       "RemoveAccess": "Retirer les accès",
+      "RemoveRights": "Supprimer les droits",
       "AreYouSureRemoveAccess": "Êtes-vous sûr de vouloir retirer les accès pour l'utilisateur {firstname} {lastname} pour ce service ?",
       "InviteUser": {
         "TitleInviteUser": "Inviter un utilisateur",
@@ -158,7 +160,8 @@
         "NumberDownload": "Téléchargements"
       },
       "Vault": "Partenaire Vault",
-      "VaultTitle": "Administration du Vault"
+      "VaultTitle": "Administration du Vault",
+      "UpdateDocument": "Mise à jour du document"
     }
   },
   "ServiceActions": {
@@ -193,6 +196,7 @@
     "Success": "Succès",
     "Update": "Editer",
     "Validate": "Valider",
+    "ValidateForm": "Valider le formulaire",
     "NotFound": "Non trouvé"
   },
   "DeleteUserDialog": {
@@ -202,7 +206,7 @@
   },
   "UpdateUserDialog": {
     "TextUpdatedUser": "Votre compte a été modifié par un administrateur. Le comportement de l'application peut donc évoluer.",
-    "TextUpdatedUserTitle":  "Compte modifié !"
+    "TextUpdatedUserTitle": "Compte modifié !"
   },
   "RemoveUserOrgDialog": {
     "TextRemoveThisUser": "Êtes-vous sûr de vouloir retirer cet utilisateur {email} de votre organisation ?"
@@ -223,6 +227,14 @@
     "AddOrganization": "Abonnement de l'organisation au service",
     "AddOrganizationDescription": "Inviter une organisation sur le service."
   },
+  "InviteUserServiceForm": {
+    "Title": "Inviter l'utilisateur au service",
+    "Description": "Définissez ici les droits d'accès. Cliquez sur Valider lorsque vous avez terminé.",
+    "Email": "Email",
+    "Organization": "Organisation",
+    "Capabilities": "Capacités",
+    "CapabilitiesPlaceholder": "Sélectionner les capacités"
+  },
   "UserActions": {
     "DeleteUser": "Supprimer l'utilisateur",
     "UpdateUser": "Mettre à jour l'utilisateur",
@@ -234,7 +246,14 @@
     "UserUpdated": "L'utilisateur {email} a été modifié",
     "UserDeleted": "L'utilisateur {email} a été supprimé",
     "SearchUserWithEmail": "Rechercher l'utilisateur par email..."
-
+  },
+  "UserForm": {
+    "Organizations": "Organisations",
+    "OrganizationsPlaceholder": "Sélectionner une organisation",
+    "Roles": "Rôles",
+    "RolesPlaceholder": "Sélectionner les rôles",
+    "Password": "Mot de passe",
+    "Email": "Courriel"
   },
   "OrganizationForm": {
     "EditOrganization": "Modifier l'organisation",
@@ -269,7 +288,12 @@
     "Delete": "Supprimer",
     "Update": "Modifier",
     "Details": "Détails",
-    "Remove": "Retirer"
+    "Remove": "Retirer",
+    "Continue": "Continuer"
+  },
+  "DialogActions": {
+    "ContinueTitle": "Progrès non sauvegardé",
+    "ContinueSentence": "Vous avez commencé à remplir le formulaire. Le fait de le fermer maintenant annulera toute progression non enregistrée. Voulez-vous continuer ?"
   },
   "RedirectPage": {
     "YourFutureHub": "Votre futur centre communautaire Filigran prend forme !",
