+++ portal-front/src/components/admin/user/user-create.tsx
@@ -1,9 +1,10 @@
-import { UserFormSheet } from '@/components/admin/user/user-form-sheet';
+import { UserForm } from '@/components/admin/user/user-form';
 import {
   userEditFormSchema,
   userFormSchema,
 } from '@/components/admin/user/user-form.schema';
 import { UserListCreateMutation } from '@/components/admin/user/user.graphql';
+import { SheetWithPreventingDialog } from '@/components/ui/sheet-with-preventing-dialog';
 import TriggerButton from '@/components/ui/trigger-button';
 import { useToast } from 'filigran-ui';
 import { useTranslations } from 'next-intl';
@@ -51,15 +52,18 @@ export const AddUser: FunctionComponent<CreateUserProps> = ({
       },
     });
   };
+
   return (
-    <UserFormSheet
-      title={t('UserActions.CreateUser')}
-      handleSubmit={handleSubmit}
-      open={openSheet}
+    <SheetWithPreventingDialog
+      title={t('UserActions.AddUser')}
       setOpen={setOpenSheet}
-      validationSchema={userFormSchema}
-      user={undefined}
-      trigger={<TriggerButton label={t('UserActions.AddUser')} />}
-    />
+      open={openSheet}
+      trigger={<TriggerButton label={t('UserActions.AddUser')} />}>
+      <UserForm
+        handleSubmit={handleSubmit}
+        user={undefined}
+        validationSchema={userFormSchema}
+      />
+    </SheetWithPreventingDialog>
   );
 };
