+++ portal-front/src/components/service/[slug]/service-slug.tsx
@@ -1,8 +1,8 @@
 import { SubscriptionDeleteMutation } from '@/components/subcription/subscription.graphql';
 
 import { Portal, portalContext } from '@/components/portal-context';
-import { ServiceSlugAddOrgaFormSheet } from '@/components/service/[slug]/service-slug-add-orga-form-sheet';
-import { ServiceSlugFormSheet } from '@/components/service/[slug]/service-slug-form-sheet';
+import { ServiceSlugAddOrgaForm } from '@/components/service/[slug]/service-slug-add-orga-form';
+import { ServiceSlugForm } from '@/components/service/[slug]/service-slug-form';
 import ServiceUserServiceSlug from '@/components/service/[slug]/service-user-service-table';
 import { ServiceByIdWithSubscriptions } from '@/components/service/service.graphql';
 import { AlertDialogComponent } from '@/components/ui/alert-dialog';
@@ -10,6 +10,7 @@ import {
   BreadcrumbNav,
   BreadcrumbNavLink,
 } from '@/components/ui/breadcrumb-nav';
+import { SheetWithPreventingDialog } from '@/components/ui/sheet-with-preventing-dialog';
 import TriggerButton from '@/components/ui/trigger-button';
 import useAdminPath from '@/hooks/useAdminPath';
 import useGranted from '@/hooks/useGranted';
@@ -195,38 +196,48 @@ const ServiceSlug: FunctionComponent<ServiceSlugProps> = ({
         <DataTableHeadBarOptions />
 
         {useAdminPath() && (
-          <ServiceSlugAddOrgaFormSheet
-            subscriptions={
-              queryData?.serviceByIdWithSubscriptions
-                ?.subscriptions as subscriptionWithUserService_fragment$data[]
-            }
-            setSelectedSubscription={setSelectedSubscription}
+          <SheetWithPreventingDialog
             open={openSheetAddOrga}
             setOpen={setOpenSheetAddOrga}
-            serviceId={serviceId}
             trigger={
               <Button variant="outline">
                 {t('Service.SubscribeOrganization')}
               </Button>
             }
-          />
+            title={t('OrganizationInServiceAction.AddOrganization')}
+            description={t(
+              'OrganizationInServiceAction.AddOrganizationDescription'
+            )}>
+            <ServiceSlugAddOrgaForm
+              subscriptions={
+                queryData?.serviceByIdWithSubscriptions
+                  ?.subscriptions as subscriptionWithUserService_fragment$data[]
+              }
+              setSelectedSubscription={setSelectedSubscription}
+              serviceId={serviceId}
+            />
+          </SheetWithPreventingDialog>
         )}
 
         {dataOrganizationsTab.length > 0 && (
-          <ServiceSlugFormSheet
+          <SheetWithPreventingDialog
             open={openSheet}
             setOpen={setOpenSheet}
-            userService={currentUser}
-            connectionId={queryData.serviceByIdWithSubscriptions?.__id ?? ''}
-            dataOrganizationsTab={dataOrganizationsTab}
-            subscription={selectedSubscription}
             trigger={
               <TriggerButton
                 onClick={() => setCurrentUser({})}
                 label={t('Service.Management.InviteUser.TitleInviteUser')}
               />
             }
-          />
+            title={t('InviteUserServiceForm.Title')}
+            description={t('InviteUserServiceForm.Description')}>
+            <ServiceSlugForm
+              userService={currentUser}
+              connectionId={queryData.serviceByIdWithSubscriptions?.__id ?? ''}
+              dataOrganizationsTab={dataOrganizationsTab}
+              subscription={selectedSubscription}
+            />
+          </SheetWithPreventingDialog>
         )}
       </div>
     </div>
